requirements.txt
# === SAMBA EXPORT CONTROL DESK ===
# Core
python-dotenv==1.1.0
pydantic==2.11.3

# Database
sqlalchemy==2.0.40

# AI
anthropic==0.52.0

# Google APIs
google-api-python-client==2.170.0
google-auth-oauthlib==1.2.1
gspread==6.1.4

# WhatsApp (Twilio)
twilio==9.6.1

# Web Scraping
requests==2.32.3
beautifulsoup4==4.13.4

# Dashboard
streamlit==1.45.1
plotly==6.1.2

# Data
pandas==2.2.3
openpyxl==3.1.5

# Scheduling & Eventing
celery==5.5.2
redis==6.2.0
schedule==1.2.2

# API (Webhook Twilio + futuro Conversational Hub)
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Utils
rich==14.0.0
click==8.2.1
PyYAML==6.0.2
tenacity==9.1.4
